Move session related classes to the sessions package
COSDK-500
