+++ core/src/main/java/com/adyen/checkout/core/internal/ui/model/CommonComponentParamsMapper.kt
@@ -11,6 +11,7 @@ package com.adyen.checkout.core.internal.ui.model
 import androidx.annotation.RestrictTo
 import com.adyen.checkout.core.analytics.internal.AnalyticsParams
 import com.adyen.checkout.core.components.CheckoutConfiguration
+import com.adyen.checkout.core.sessions.internal.model.SessionParams
 import java.util.Locale
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
