+++ core/src/main/java/com/adyen/checkout/core/internal/ui/model/DropInOverrideParams.kt
@@ -10,6 +10,7 @@ package com.adyen.checkout.core.internal.ui.model
 
 import androidx.annotation.RestrictTo
 import com.adyen.checkout.core.components.data.model.Amount
+import com.adyen.checkout.core.sessions.internal.model.SessionParams
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class DropInOverrideParams(
