+++ core/src/main/java/com/adyen/checkout/core/sessions/internal/SessionsComponentEventHandler.kt
@@ -6,11 +6,13 @@
  * Created by ozgur on 16/5/2025.
  */
 
-package com.adyen.checkout.core.internal
+package com.adyen.checkout.core.sessions.internal
 
 import com.adyen.checkout.core.components.CheckoutCallback
 import com.adyen.checkout.core.components.CheckoutResult
 import com.adyen.checkout.core.components.paymentmethod.PaymentComponentState
+import com.adyen.checkout.core.internal.ComponentEventHandler
+import com.adyen.checkout.core.internal.PaymentComponentEvent
 import com.adyen.checkout.core.sessions.SessionInteractor
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
